# src/main/java/io/electrum/prepaidutility/api/MeterLookupsResource.java
package io.electrum.prepaidutility.api;

import javax.servlet.http.HttpServletRequest;
import javax.ws.rs.Consumes;
import javax.ws.rs.POST;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;
import javax.ws.rs.container.AsyncResponse;
import javax.ws.rs.container.Suspended;
import javax.ws.rs.core.Context;
import javax.ws.rs.core.HttpHeaders;
import javax.ws.rs.core.Request;
import javax.ws.rs.core.SecurityContext;
import javax.ws.rs.core.UriInfo;

import io.electrum.prepaidutility.model.MeterLookupRequest;
import io.electrum.prepaidutility.model.MeterLookupResponse;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import io.swagger.annotations.Authorization;

@Path("/meterLookups")

@Api(description = "the meterLookups API", authorizations = { @Authorization("httpBasic") })
@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2016-08-26T07:31:58.129Z")
public abstract class MeterLookupsResource {

   protected abstract IMeterLookupsResource getResourceImplementation();

   @POST
   @Path("/{lookupId}")
   @Consumes({ "application/json" })
   @Produces({ "application/json" })
   @ApiOperation(value = "Request information about a specified meter", notes = "Request information about a specified meter, including customer and utility details. This resource is used to verify that a meter number is valid and registered with the correct details. It also confirms whether the meter is recognised by a provider and that tokens can be issued against it.")
   @ApiResponses(value = { @ApiResponse(code = 201, message = "Created", response = MeterLookupResponse.class),
         @ApiResponse(code = 400, message = "Bad request", response = MeterLookupResponse.class),
         @ApiResponse(code = 500, message = "Internal Server Error", response = MeterLookupResponse.class),
         @ApiResponse(code = 503, message = "Service Unavailable", response = MeterLookupResponse.class),
         @ApiResponse(code = 504, message = "Gateway Timeout", response = MeterLookupResponse.class) })
   public void createMeterLookup(
         @ApiParam(value = "The randomly generated UUID of this request.", required = true) @PathParam("lookupId") String lookupId,
         @ApiParam(value = "A meter information lookup request.", required = true) MeterLookupRequest body,
         @Context SecurityContext securityContext,
         @Suspended AsyncResponse asyncResponse,
         @Context Request request,
         @Context HttpServletRequest httpServletRequest,
         @Context HttpHeaders httpHeaders,
         @Context UriInfo uriInfo) {

      getResourceImplementation().createMeterLookup(
            lookupId,
            body,
            securityContext,
            asyncResponse,
            request,
            httpServletRequest,
            httpHeaders,
            uriInfo);
   }
}